Fix build warnings
